+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## UNRELEASED
 
+### IMPROVEMENTS
+
+* Split workflow execution unit to step in order to allow a unique workflow to be executed by multiple Yorc instances. ([GH-93](https://github.com/ystia/yorc/issues/93))
+* Make the run step of a Job execution asynchronous not to block a worker during the duration of the job. ([GH-85](https://github.com/ystia/yorc/issues/85))
+
+### BUG FIXES
+
+* inputs are not injected into Slurm (srun) jobs ([GH-161](https://github.com/ystia/yorc/issues/161))
+
 ## 3.1.0-M3 (September 14, 2018)
 
 ### IMPROVEMENTS